refactor: separate design, logic, demo of managing data
